backend/requirements.txt
# FastAPI & server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9   # PostgreSQL driver (Supabase)
aiosqlite==0.20.0         # SQLite async driver (local dev)

# ML / Data (Updated for Python 3.13 compatibility)
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0             # model serialization

# Data generation
faker>=24.0.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.7.1
python-dateutil==2.9.0
httpx==0.27.0             # for internal service calls

# Supabase client
supabase>=2.0.0

# LLM Providers
groq>=0.4.0
